+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 2.6, 2024-06-23
 ---------------
+- add option to delete from disk in files list
 - fix missing creatures in Horn of the Abyss (pull request #13)
 - fix Halfling having duplicate name Halfing (pull request #13)
 - fix Halfling Grenadier being named Halfing Grenadier (pull request #13)